refactor: promote shared assembly viewmodelbuilders to shared (#1619)
