+++ packages/shared/viewModelBuilders/viewModelBuilders.tsx
@@ -9,6 +9,7 @@ import { Stack } from "@databiosphere/findable-ui/lib/components/common/Stack/st
 import { Link } from "@databiosphere/findable-ui/lib/components/Links/components/Link/link";
 import { type BasicCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/BasicCell/basicCell";
 import { type ChipCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/ChipCell/chipCell";
+import { type NTagCell } from "@databiosphere/findable-ui/lib/components/Table/components/TableCell/components/NTagCell/nTagCell";
 import { CHIP_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/chip";
 import { replaceParameters } from "@databiosphere/findable-ui/lib/utils/replaceParameters";
 import type {
@@ -41,6 +42,19 @@ import {
 } from "./constants";
 
 /**
+ * Build props for the accession cell.
+ * @param entity - Entity with an accession.
+ * @returns Props to be used for the cell.
+ */
+export const buildAccession = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.accession,
+  };
+};
+
+/**
  * Build props for the genome analysis cell.
  * @param entity - Entity with an accession, ncbiTaxonomyId, and ucscBrowserUrl.
  * @returns Props to be used for the AnalyzeGenome component.
@@ -71,6 +85,34 @@ export const buildAnalyzeGenome = (
         )}`,
       },
     ],
+  };
+};
+
+/**
+ * Build props for the annotation status cell.
+ * @param entity - Entity with an annotationStatus property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildAnnotationStatus = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.annotationStatus,
+  };
+};
+
+/**
+ * Build props for the assemblies cell.
+ * @param entity - Entity with a required assemblyCount property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildAssemblyCount = (
+  // `assemblyCount` is optional on OrganismContract; require it here so callers
+  // can't pass a projection without it and silently render an empty value.
+  entity: OrganismContract & { assemblyCount: number }
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: formatNumber(entity.assemblyCount),
   };
 };
 
@@ -156,6 +198,72 @@ export const buildAssemblyResources = (
 };
 
 /**
+ * Build props for the taxonomic group cell of an assembly.
+ * @param entity - Assembly with a taxonomicGroup property.
+ * @returns Props for the NTagCell component.
+ */
+export const buildAssemblyTaxonomicGroup = (
+  entity: AssemblyContract
+): ComponentProps<typeof NTagCell> => {
+  return {
+    label: "taxonomic groups",
+    values: entity.taxonomicGroup,
+  };
+};
+
+/**
+ * Build props for the chromosomes cell.
+ * @param entity - Entity with a chromosomes property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildChromosomes = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: formatNumber(entity.chromosomes),
+  };
+};
+
+/**
+ * Build props for the coverage cell.
+ * @param entity - Entity with a coverage property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildCoverage = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.coverage,
+  };
+};
+
+/**
+ * Build props for the GC% cell.
+ * @param entity - Entity with a gcPercent property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildGcPercent = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.gcPercent,
+  };
+};
+
+/**
+ * Build props for the strain cell.
+ * @param entity - Entity with a strainName and taxonomicLevelStrain property.
+ * @returns Props to be used for the BasicCell component.
+ */
+export const buildGenomeTaxonomicLevelStrain = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: getGenomeStrainText(entity),
+  };
+};
+
+/**
  * Build the taxonomic group tag, or null when there is no group.
  * @param taxonomicGroup - Taxonomic group values.
  * @returns Group tag, or null.
@@ -187,6 +295,19 @@ export const buildIsRef = (
 };
 
 /**
+ * Build props for the length cell.
+ * @param entity - Entity with a length property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildLength = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: formatNumber(entity.length),
+  };
+};
+
+/**
  * Build props for the level cell — a tiered bar indicator plus the level label.
  * @param entity - Entity with a level property.
  * @returns Props for the LevelCell component.
@@ -197,6 +318,20 @@ export const buildLevel = (
   return {
     filledCount: LEVEL_FILLED_COUNT[entity.level] ?? 0,
     label: LEVEL_LABEL[entity.level] ?? entity.level,
+  };
+};
+
+/**
+ * Build props for the assembly taxonomy IDs cell.
+ * @param entity - Entity with a assemblyTaxonomyIds property.
+ * @returns Props for the NTagCell component.
+ */
+export const buildOrganismAssemblyTaxonomyIds = (
+  entity: OrganismContract
+): ComponentProps<typeof NTagCell> => {
+  return {
+    label: "taxonomy IDs",
+    values: entity.assemblyTaxonomyIds ?? [],
   };
 };
 
@@ -245,6 +380,20 @@ export const buildOrganismDetails = (
 };
 
 /**
+ * Build props for the taxonomic group cell of an organism.
+ * @param entity - Organism with an optional taxonomicGroup property.
+ * @returns Props for the NTagCell component.
+ */
+export const buildOrganismTaxonomicGroup = (
+  entity: OrganismContract
+): ComponentProps<typeof NTagCell> => {
+  return {
+    label: "taxonomic groups",
+    values: entity.taxonomicGroup ?? [],
+  };
+};
+
+/**
  * Build props for the release date cell, displaying the release year.
  * @param entity - Entity with a releaseDate property.
  * @returns Props for the BasicCell component.
@@ -272,6 +421,156 @@ export const buildReleaseDateTooltip = (
     title: entity.releaseDate
       ? formatDate(parseISO(entity.releaseDate))
       : undefined,
+  };
+};
+
+/**
+ * Build props for the scaffold count cell.
+ * @param entity - Entity with a scaffoldCount property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildScaffoldCount = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: formatNumber(entity.scaffoldCount),
+  };
+};
+
+/**
+ * Build props for the scaffold L50 cell.
+ * @param entity - Entity with a scaffoldL50 property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildScaffoldL50 = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: formatNumber(entity.scaffoldL50),
+  };
+};
+
+/**
+ * Build props for the scaffold N50 cell.
+ * @param entity - Entity with a scaffoldN50 property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildScaffoldN50 = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: formatNumber(entity.scaffoldN50),
+  };
+};
+
+/**
+ * Build props for the class cell.
+ * @param entity - Entity with a taxonomicLevelClass property.
+ * @param entity.taxonomicLevelClass - Taxonomic class.
+ * @returns Props to be used for the cell.
+ */
+export const buildTaxonomicLevelClass = (entity: {
+  taxonomicLevelClass: string;
+}): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.taxonomicLevelClass,
+  };
+};
+
+/**
+ * Build props for the domain cell.
+ * @param entity - Entity with a taxonomicLevelDomain property.
+ * @param entity.taxonomicLevelDomain - Taxonomic domain.
+ * @returns Props to be used for the cell.
+ */
+export const buildTaxonomicLevelDomain = (entity: {
+  taxonomicLevelDomain: string;
+}): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.taxonomicLevelDomain,
+  };
+};
+
+/**
+ * Build props for the family cell.
+ * @param entity - Entity with a taxonomicLevelFamily property.
+ * @param entity.taxonomicLevelFamily - Taxonomic family.
+ * @returns Props to be used for the cell.
+ */
+export const buildTaxonomicLevelFamily = (entity: {
+  taxonomicLevelFamily: string;
+}): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.taxonomicLevelFamily,
+  };
+};
+
+/**
+ * Build props for the genus cell.
+ * @param entity - Entity with a taxonomicLevelGenus property.
+ * @param entity.taxonomicLevelGenus - Taxonomic genus.
+ * @returns Props to be used for the cell.
+ */
+export const buildTaxonomicLevelGenus = (entity: {
+  taxonomicLevelGenus: string;
+}): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.taxonomicLevelGenus,
+  };
+};
+
+/**
+ * Build props for the kingdom cell.
+ * @param entity - Entity with a taxonomicLevelKingdom property.
+ * @param entity.taxonomicLevelKingdom - Taxonomic kingdom.
+ * @returns Props to be used for the cell.
+ */
+export const buildTaxonomicLevelKingdom = (entity: {
+  taxonomicLevelKingdom: string;
+}): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.taxonomicLevelKingdom,
+  };
+};
+
+/**
+ * Build props for the order cell.
+ * @param entity - Entity with a taxonomicLevelOrder property.
+ * @param entity.taxonomicLevelOrder - Taxonomic order.
+ * @returns Props to be used for the cell.
+ */
+export const buildTaxonomicLevelOrder = (entity: {
+  taxonomicLevelOrder: string;
+}): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.taxonomicLevelOrder,
+  };
+};
+
+/**
+ * Build props for the phylum cell.
+ * @param entity - Entity with a taxonomicLevelPhylum property.
+ * @param entity.taxonomicLevelPhylum - Taxonomic phylum.
+ * @returns Props to be used for the cell.
+ */
+export const buildTaxonomicLevelPhylum = (entity: {
+  taxonomicLevelPhylum: string;
+}): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.taxonomicLevelPhylum,
+  };
+};
+
+/**
+ * Build props for the taxonomy ID cell.
+ * @param entity - Entity with a ncbiTaxonomyId property.
+ * @returns Props for the BasicCell component.
+ */
+export const buildTaxonomyId = (
+  entity: AssemblyContract
+): ComponentProps<typeof BasicCell> => {
+  return {
+    value: entity.ncbiTaxonomyId,
   };
 };
