+++ site-config/brc-analytics/local/index/genomeEntityConfig.ts
@@ -17,11 +17,30 @@ import { SpeciesCell } from "@repo/shared/components/Table/components/TableCell/
 import { Tooltip } from "@repo/shared/components/Tooltip/tooltip";
 import { type AppEntityConfig } from "@repo/shared/config/types";
 import {
+  buildAccession,
   buildAnalyzeGenome,
+  buildAnnotationStatus,
+  buildAssemblyTaxonomicGroup,
+  buildChromosomes,
+  buildCoverage,
+  buildGcPercent,
+  buildGenomeTaxonomicLevelStrain,
   buildIsRef,
+  buildLength,
   buildLevel,
   buildReleaseDate,
   buildReleaseDateTooltip,
+  buildScaffoldCount,
+  buildScaffoldL50,
+  buildScaffoldN50,
+  buildTaxonomicLevelClass,
+  buildTaxonomicLevelDomain,
+  buildTaxonomicLevelFamily,
+  buildTaxonomicLevelGenus,
+  buildTaxonomicLevelKingdom,
+  buildTaxonomicLevelOrder,
+  buildTaxonomicLevelPhylum,
+  buildTaxonomyId,
 } from "@repo/shared/viewModelBuilders/viewModelBuilders";
 import {
   BRC_DATA_CATALOG_CATEGORY_KEY,
@@ -156,7 +175,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomicLevelDomain,
+          viewBuilder: buildTaxonomicLevelDomain,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_DOMAIN,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.TAXONOMIC_LEVEL_DOMAIN,
@@ -174,7 +193,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomicLevelKingdom,
+          viewBuilder: buildTaxonomicLevelKingdom,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_KINGDOM,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.TAXONOMIC_LEVEL_KINGDOM,
@@ -183,7 +202,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomicLevelPhylum,
+          viewBuilder: buildTaxonomicLevelPhylum,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_PHYLUM,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.TAXONOMIC_LEVEL_PHYLUM,
@@ -192,7 +211,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomicLevelClass,
+          viewBuilder: buildTaxonomicLevelClass,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_CLASS,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.TAXONOMIC_LEVEL_CLASS,
@@ -201,7 +220,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomicLevelOrder,
+          viewBuilder: buildTaxonomicLevelOrder,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_ORDER,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.TAXONOMIC_LEVEL_ORDER,
@@ -210,7 +229,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomicLevelFamily,
+          viewBuilder: buildTaxonomicLevelFamily,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_FAMILY,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.TAXONOMIC_LEVEL_FAMILY,
@@ -219,7 +238,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomicLevelGenus,
+          viewBuilder: buildTaxonomicLevelGenus,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_GENUS,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.TAXONOMIC_LEVEL_GENUS,
@@ -238,7 +257,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildAccession,
+          viewBuilder: buildAccession,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.ACCESSION,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.ACCESSION,
@@ -267,7 +286,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildGenomeTaxonomicLevelStrain,
+          viewBuilder: buildGenomeTaxonomicLevelStrain,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         enableHiding: false,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_LEVEL_STRAIN,
@@ -306,7 +325,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildTaxonomyId,
+          viewBuilder: buildTaxonomyId,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         enableHiding: false,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMY_ID,
@@ -316,7 +335,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: NTagCell,
-          viewBuilder: V.buildAssemblyTaxonomicGroup,
+          viewBuilder: buildAssemblyTaxonomicGroup,
         } as ComponentConfig<typeof NTagCell, BRCDataCatalogGenome>,
         enableHiding: false,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.TAXONOMIC_GROUP,
@@ -359,7 +378,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildChromosomes,
+          viewBuilder: buildChromosomes,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.CHROMOSOMES,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.CHROMOSOMES,
@@ -368,7 +387,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildLength,
+          viewBuilder: buildLength,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.LENGTH,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.LENGTH,
@@ -377,7 +396,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildScaffoldCount,
+          viewBuilder: buildScaffoldCount,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.SCAFFOLD_COUNT,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.SCAFFOLD_COUNT,
@@ -386,7 +405,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildScaffoldN50,
+          viewBuilder: buildScaffoldN50,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.SCAFFOLD_N50,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.SCAFFOLD_N50,
@@ -395,7 +414,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildScaffoldL50,
+          viewBuilder: buildScaffoldL50,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.SCAFFOLD_L50,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.SCAFFOLD_L50,
@@ -404,7 +423,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildCoverage,
+          viewBuilder: buildCoverage,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.COVERAGE,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.COVERAGE,
@@ -413,7 +432,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildGcPercent,
+          viewBuilder: buildGcPercent,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.GC_PERCENT,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.GC_PERCENT,
@@ -422,7 +441,7 @@ export const genomeEntityConfig: AppEntityConfig<BRCDataCatalogGenome> = {
       {
         componentConfig: {
           component: BasicCell,
-          viewBuilder: V.buildAnnotationStatus,
+          viewBuilder: buildAnnotationStatus,
         } as ComponentConfig<typeof BasicCell, BRCDataCatalogGenome>,
         header: BRC_DATA_CATALOG_CATEGORY_LABEL.ANNOTATION_STATUS,
         id: BRC_DATA_CATALOG_CATEGORY_KEY.ANNOTATION_STATUS,
